Move CacheFactory from cache to infrastructure.core
