+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/infrastructure/core/EhcacheConfiguration.java
@@ -24,7 +24,6 @@ package com.tesshu.jpsonic.infrastructure.core;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import com.tesshu.jpsonic.cache.CacheFactory;
 import net.sf.ehcache.CacheManager;
 import net.sf.ehcache.Ehcache;
 import org.slf4j.Logger;
