+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/infrastructure/core/CacheFactoryTest.java
@@ -1,0 +1,42 @@
+/*
+ * This file is part of Jpsonic.
+ *
+ * Jpsonic is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Jpsonic is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * (C) 2026 tesshucom
+ */
+
+package com.tesshu.jpsonic.infrastructure.core;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+
+import org.junit.jupiter.api.Test;
+
+@NeedsHome
+class CacheFactoryTest {
+
+    @Test
+    void testGetCache() {
+        CacheFactory cacheFactory = new CacheFactory();
+        cacheFactory.afterPropertiesSet();
+
+        assertNull(cacheFactory.getCache("mediaFile"));
+
+        assertNotNull(cacheFactory.getCache("mediaFileMemoryCache"));
+        assertNotNull(cacheFactory.getCache("genreCache"));
+        assertNotNull(cacheFactory.getCache("randomCache"));
+        assertNotNull(cacheFactory.getCache("fontCache"));
+    }
+}
